+++ TildarMen/src/main/java/kz/tildarmen/TildarMen/controller/JobController.java
@@ -1,13 +1,16 @@
 package kz.tildarmen.TildarMen.controller;
 
 import kz.tildarmen.TildarMen.dto.JobDto;
+import kz.tildarmen.TildarMen.model.User;
 import kz.tildarmen.TildarMen.requests.SearchJobsRequest;
 import kz.tildarmen.TildarMen.response.ApiResponse;
+import kz.tildarmen.TildarMen.services.AuthService;
 import kz.tildarmen.TildarMen.services.JobService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -18,6 +21,7 @@ import java.util.List;
 public class JobController {
 
     private final JobService jobService;
+    private final AuthService authService;
 
 
     @GetMapping("/all")
@@ -37,11 +41,17 @@ public class JobController {
 
     @PreAuthorize("hasAnyAuthority('EMPLOYER')")
     @PostMapping("/{employerId}/add")
-    public ResponseEntity<ApiResponse> addJob(@PathVariable Long employerId, @RequestBody JobDto job) {
+    public ResponseEntity<ApiResponse> addJob(@PathVariable Long employerId, @RequestBody JobDto job,
+                                              @AuthenticationPrincipal User userDetails) {
         try {
+            authService.checkPermission(userDetails, employerId);
             JobDto newJob = jobService.addJob(employerId, job);
             return ResponseEntity.ok(new ApiResponse("Success", newJob));
-        } catch (Exception e) {
+        } catch (SecurityException e){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(new ApiResponse("Forbidden", e.getMessage()));
+        }
+        catch (Exception e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiResponse("Error", e.getMessage()));
         }
     }
@@ -50,22 +60,34 @@ public class JobController {
     @PutMapping("/{employerId}/update/{jobId}")
     public ResponseEntity<ApiResponse> updateJob(@PathVariable Long employerId,
                                                  @PathVariable Long jobId,
-                                                 @RequestBody JobDto job) {
+                                                 @RequestBody JobDto job,
+                                                 @AuthenticationPrincipal User userDetails) {
         try {
+            authService.checkPermission(userDetails, employerId);
             JobDto newJob = jobService.updateJobById(employerId, jobId, job);
             return ResponseEntity.ok(new ApiResponse("Success", newJob));
-        } catch (Exception e) {
+        } catch (SecurityException e){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(new ApiResponse("Forbidden", e.getMessage()));
+        }
+        catch (Exception e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiResponse("Error", e.getMessage()));
         }
     }
 
     @PreAuthorize("hasAnyAuthority('EMPLOYER')")
     @DeleteMapping("/{employerId}/delete/{jobId}")
-    public ResponseEntity<ApiResponse> deleteJob(@PathVariable Long employerId, @PathVariable Long jobId) {
+    public ResponseEntity<ApiResponse> deleteJob(@PathVariable Long employerId, @PathVariable Long jobId,
+                                                 @AuthenticationPrincipal User userDetails) {
         try {
+            authService.checkPermission(userDetails, employerId);
             jobService.deleteJobByEmployerId(employerId, jobId);
             return ResponseEntity.ok(new ApiResponse("Delete Success", null));
-        } catch (Exception e) {
+        } catch (SecurityException e){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(new ApiResponse("Forbidden", e.getMessage()));
+        }
+        catch (Exception e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ApiResponse("Error", e.getMessage()));
         }
     }
